new conditions added

+++ public/uploadPdfPage.js
@@ -1,24 +1,8 @@
 import showDataPage from './showDataPage.js';
 import loginPage from './loginPage.js';
-import { notify,popup,confirmationPopup } from './library.js';
+import { notify, popup, confirmationPopup } from './library.js';
 
 const uploadPdfPage = () => {
-    popup({
-       state: true,
-       content: confirmationPopup({
-        title: 'Upload PDF',
-        message: 'Upload a PDF file to parse data from it',
-        callback: ()=>{
-            console.log('ashik lol')
-            popup({state:false});
-        },
-       }),
-       options:{
-        backDrop: true,
-        removeButton: true,
-        backDropColor: 'rgba(0,0,0,0.75)',
-       }
-    })
     const body = document.querySelector('body');
     const uploadPdfPage = document.createElement('div');
     uploadPdfPage.classList.add('upload-pdf-page');
@@ -36,21 +20,62 @@ const uploadPdfPage = () => {
     const verifyUploadPdf = async () => {
         const formData = new FormData();
         formData.append('pdf', uploadPdfFormInput.files[0]);
-        const response = await fetch('/api/datas/parse-pdf', {
-            method: 'POST',
-            body: formData
+        const responseData = await fetch(`/api/datas/is-pdf-exists/${uploadPdfFormInput.files[0].name}`, {
+            method: 'GET',
         });
-        const data = await response.json();
-        if (response.status === 401) {
-            localStorage.setItem('currentPage', '0');
+        const jsonData = await responseData.json();
+        if (jsonData.isPdfExists) {
+            popup({
+                state: true,
+                content: confirmationPopup({
+                    title: 'This PDF is already uploaded',
+                    message: 'Do you want to upload it again?',
+                    callback: async () => {
+                        popup({ state: false });
+                        const response = await fetch('/api/datas/parse-pdf', {
+                            method: 'POST',
+                            body: formData
+                        });
+                        const data = await response.json();
+                        if (response.status === 401) {
+                            localStorage.setItem('currentPage', '0');
+                            body.removeChild(uploadPdfPage);
+                            loginPage();
+                        } else if (response.status !== 200) {
+                            notify({ data, type: 'danger' })
+                            return;
+                        }
+                        body.removeChild(uploadPdfPage);
+                        showDataPage(data);
+                    },
+                    negativeCallBack: async () => {
+                        popup({ state: false });
+                        uploadPdfFormInput.value = '';
+                    }
+                }),
+                options: {
+                    backDrop: false,
+                    removeButton: false,
+                    backDropColor: 'rgba(0,0,0,0.75)',
+                }
+            });
+        } else {
+            const response = await fetch('/api/datas/parse-pdf', {
+                method: 'POST',
+                body: formData
+            });
+            const data = await response.json();
+            if (response.status === 401) {
+                localStorage.setItem('currentPage', '0');
+                body.removeChild(uploadPdfPage);
+                loginPage();
+            } else if (response.status !== 200) {
+                notify({ data, type: 'danger' })
+                return;
+            }
             body.removeChild(uploadPdfPage);
-            loginPage();
-        } else if (response.status !== 200) {
-            notify({ data, type: 'danger' })
-            return;
+            showDataPage(data);
         }
-        body.removeChild(uploadPdfPage);
-        showDataPage(data);
     }
     const uploadIcon = document.createElement('label');
     uploadIcon.classList.add('upload-label');
